fixed external resource + encodeURI

+++ server/api/reset.js
@@ -3,5 +3,6 @@ import process from "process"
 
 export default defineEventHandler(async()=>{
     const sourceDirectory = process.cwd() + "/static/archive"
+    if(!fs.existsSync(sourceDirectory)) return;
     await fs.rmdirSync(sourceDirectory, {recursive: true, force: true})
 })
